requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
rapidfuzz>=3.0.0
pymongo>=4.6.0
dnspython>=2.4.0
flask-cors>=4.0.0

